+++ Sources/VComponents/Components/Modals/Alert/VAlertButtonConvertible.swift
@@ -9,7 +9,6 @@ import SwiftUI
 
 // MARK: - V Alert Button Convertible
 /// Type that allows for conversion to `VAlertButtonProtocol`.
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)
@@ -18,7 +17,6 @@ public protocol VAlertButtonConvertible {
     func toButtons() -> [any VAlertButtonProtocol]
 }
 
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)
@@ -26,7 +24,6 @@ extension Array: VAlertButtonConvertible where Element == any VAlertButtonProtoc
     public func toButtons() -> [any VAlertButtonProtocol] { self }
 }
 
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)