+++ Sources/VComponents/Components/Modals/Alert/VAlertExtension.swift
@@ -9,7 +9,6 @@ import SwiftUI
 import VCore
 
 // MARK: - Bool
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)
@@ -116,7 +115,6 @@ extension View {
 }
 
 // MARK: - Item
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)
@@ -252,7 +250,6 @@ extension View {
 }
 
 // MARK: - Error
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)