+++ CHANGELOG.md
@@ -40,6 +40,7 @@ VSideBar
 VAlert
 
 - API is overhauled
+- Alert now supports `macOS`
 - Modal no longer requires `ColorScheme` as parameter when it's overridden in the root
 - `roundedCorners` and `cornerRadius` are replaced with `cornerRadii`